Use double instead of float as default

+++ Dice/Expressions/NumberExpression.cs
@@ -4,9 +4,9 @@ namespace Wgaffa.DMToolkit.Expressions
 {
     public class NumberExpression : IExpression
     {
-        public float Value { get; }
+        public double Value { get; }
 
-        public NumberExpression(float value)
+        public NumberExpression(double value)
         {
             Value = value;
         }
